consolidate _meta_value_to_json
