+++ WDL/Expr.py
@@ -1283,3 +1283,24 @@ class Apply(Base):
         f = getattr(stdlib, self.function_name, None)
         assert isinstance(f, StdLib.Function)
         return f(self, env, stdlib)
+
+
+def _meta_value_to_json(v: Any) -> Any:
+    if isinstance(v, Int):
+        return v.value
+    if isinstance(v, Float):
+        return v.value
+    if isinstance(v, Boolean):
+        return v.value
+    if isinstance(v, Null):
+        return None
+    if isinstance(v, String):
+        literal = v.literal
+        if isinstance(literal, Value.String):
+            return literal.value
+        return str(v)
+    if isinstance(v, dict):
+        return {k: _meta_value_to_json(vv) for k, vv in v.items()}
+    if isinstance(v, list):
+        return [_meta_value_to_json(vv) for vv in v]
+    return v
